Wrap server binary in a CLI
This is the first step towards combining the client and server into one
binary.
