+++ server/server.go
@@ -1,0 +1,246 @@
+package server
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+	"net/http"
+
+	raven "github.com/getsentry/raven-go"
+	"github.com/gocardless/draupnir/auth"
+	"github.com/gocardless/draupnir/exec"
+	"github.com/gocardless/draupnir/routes"
+	"github.com/gocardless/draupnir/routes/chain"
+	"github.com/gocardless/draupnir/server/config"
+	"github.com/gocardless/draupnir/store"
+	"github.com/gocardless/draupnir/version"
+	"github.com/gorilla/mux"
+	rungroup "github.com/oklog/run"
+	"github.com/pkg/errors"
+	"github.com/prometheus/common/log"
+	"golang.org/x/oauth2"
+)
+
+// ConfigFilePath is the expected path of the server configuration file
+const ConfigFilePath = "/etc/draupnir/config.toml"
+
+// Run starts the draupnir server
+func Run(logger log.Logger) {
+	logger.With("config", ConfigFilePath).Info("Loading config file")
+	cfg, err := config.Load(ConfigFilePath)
+	if err != nil {
+		logger.With("error", err.Error()).Fatal("Could not load configuration")
+	}
+	logger.Info("Configuration successfully loaded")
+
+	logger = log.With("environment", cfg.Environment)
+
+	oauthConfig := createOauthConfig(cfg.OAuthConfig)
+	authenticator := createAuthenticator(cfg, oauthConfig)
+	executor := createExecutor(cfg)
+
+	db := connectToDatabase(cfg, logger)
+	imageStore := createImageStore(db)
+	instanceStore := createInstanceStore(db)
+
+	imageRouteSet := routes.Images{
+		ImageStore:    imageStore,
+		InstanceStore: instanceStore,
+		Executor:      executor,
+	}
+
+	instanceRouteSet := routes.Instances{
+		InstanceStore: instanceStore,
+		ImageStore:    imageStore,
+		Executor:      executor,
+	}
+
+	accessTokenRouteSet := routes.AccessTokens{
+		Callbacks: make(map[string]chan routes.OAuthCallback),
+		Client:    &oauthConfig,
+	}
+
+	router := mux.NewRouter()
+
+	// Every request will be logged, and any error raised in serving the request
+	// will also be logged.
+	rootHandler := chain.
+		New(routes.NewErrorHandler(logger)).
+		Add(routes.NewRequestLogger(logger))
+
+	// If Sentry is available, attach the Sentry middleware
+	// This will report all errors to Sentry
+	if cfg.SentryDsn != "" {
+		sentryClient, err := raven.New(cfg.SentryDsn)
+		if err != nil {
+			logger.With("error", err.Error()).Fatal("Could not initialise sentry-raven client")
+		}
+
+		rootHandler = rootHandler.
+			Add(routes.NewSentryReporter(sentryClient))
+	}
+
+	// Healthcheck
+	// We don't enforce a particulate API version on this route, because it should
+	// be easy to hit to monitor the health of the system.
+	router.Methods("GET").Path("/health_check").HandlerFunc(
+		rootHandler.
+			Add(routes.WithVersion).
+			Add(routes.AsJSON).
+			Resolve(routes.HealthCheck),
+	)
+
+	// OAuth
+	// These routes are a bit special, because they don't accept or return JSON.
+	// They're intended to be used through a web browser.
+	router.Methods("GET").Path("/authenticate").HandlerFunc(
+		rootHandler.
+			Resolve(accessTokenRouteSet.Authenticate),
+	)
+
+	router.Methods("GET").Path("/oauth_callback").HandlerFunc(
+		rootHandler.
+			Add(routes.OauthErrorRenderer).
+			Resolve(accessTokenRouteSet.Callback),
+	)
+
+	// Core API routes
+	// These routes all accept and return JSON, and will enforce that the client
+	// sends a compatible API version header.
+	defaultChain := rootHandler.
+		Add(routes.DefaultErrorRenderer).
+		Add(routes.WithVersion).
+		Add(routes.AsJSON).
+		Add(routes.CheckAPIVersion(version.Version)).
+		Add(routes.Authenticate(authenticator))
+
+	// Access Tokens
+	// This route is hit before the user is authenticated, so we don't use the
+	// Authenticate middleware
+	router.Methods("POST").Path("/access_tokens").HandlerFunc(
+		rootHandler.
+			Add(middleware.DefaultErrorRenderer).
+			Add(middleware.WithVersion).
+			Add(middleware.AsJSON).
+			Add(middleware.CheckAPIVersion(version.Version)).
+			Resolve(accessTokenRouteSet.Create),
+	)
+
+	// Images
+	router.Methods("GET").Path("/images").HandlerFunc(
+		defaultChain.Resolve(imageRouteSet.List),
+	)
+
+	router.Methods("POST").Path("/images").HandlerFunc(
+		defaultChain.Resolve(imageRouteSet.Create),
+	)
+
+	router.Methods("GET").Path("/images/{id}").HandlerFunc(
+		defaultChain.Resolve(imageRouteSet.Get),
+	)
+
+	router.Methods("POST").Path("/images/{id}/done").HandlerFunc(
+		defaultChain.Resolve(imageRouteSet.Done),
+	)
+
+	router.Methods("DELETE").Path("/images/{id}").HandlerFunc(
+		defaultChain.Resolve(imageRouteSet.Destroy),
+	)
+
+	// Instances
+	router.Methods("GET").Path("/instances").HandlerFunc(
+		defaultChain.Resolve(instanceRouteSet.List),
+	)
+
+	router.Methods("POST").Path("/instances").HandlerFunc(
+		defaultChain.Resolve(instanceRouteSet.Create),
+	)
+
+	router.Methods("GET").Path("/instances/{id}").HandlerFunc(
+		defaultChain.Resolve(instanceRouteSet.Get),
+	)
+
+	router.Methods("DELETE").Path("/instances/{id}").HandlerFunc(
+		defaultChain.Resolve(instanceRouteSet.Destroy),
+	)
+
+	router.Methods("DELETE").Path("/instances/{id}").HandlerFunc(
+		defaultChain.Resolve(instanceRouteSet.Destroy),
+	)
+
+	var g rungroup.Group
+
+	// The default server for draupnir which will listen on TLS
+	server := http.Server{
+		Addr:    fmt.Sprintf(":%d", cfg.HTTPConfig.Port),
+		Handler: router,
+	}
+
+	g.Add(
+		func() error {
+			return server.ListenAndServeTLS(cfg.HTTPConfig.TLSCertificatePath, cfg.HTTPConfig.TLSPrivateKeyPath)
+		},
+		func(error) { server.Shutdown(context.Background()) },
+	)
+
+	// We then listen for insecure connections on localhost, allowing connections from
+	// within the host without requiring the user to explicitly ignore certificates.
+	serverInsecure := http.Server{
+		Addr:    fmt.Sprintf("127.0.0.1:%d", cfg.HTTPConfig.InsecurePort),
+		Handler: router,
+	}
+
+	g.Add(
+		func() error { return serverInsecure.ListenAndServe() },
+		func(error) { serverInsecure.Shutdown(context.Background()) },
+	)
+
+	if err := g.Run(); err != nil {
+		logger.Fatal(errors.Wrap(err, "could not start HTTP servers"))
+	}
+}
+
+func connectToDatabase(c config.Config, logger log.Logger) *sql.DB {
+	db, err := sql.Open("postgres", c.DatabaseURL)
+	if err != nil {
+		logger.With("error", err.Error()).Fatal("Could not connect to database")
+	}
+	return db
+}
+
+func createOauthConfig(c config.OAuthConfig) oauth2.Config {
+	return oauth2.Config{
+		ClientID:     c.ClientID,
+		ClientSecret: c.ClientSecret,
+		Scopes:       []string{"https://www.googleapis.com/auth/userinfo.email"},
+		Endpoint: oauth2.Endpoint{
+			AuthURL:  "https://accounts.google.com/o/oauth2/v2/auth",
+			TokenURL: "https://www.googleapis.com/oauth2/v4/token",
+		},
+		RedirectURL: c.RedirectURL,
+	}
+}
+
+func createAuthenticator(c config.Config, oauthConfig oauth2.Config) auth.Authenticator {
+	authenticator := auth.GoogleAuthenticator{
+		OAuthClient:            auth.GoogleOAuthClient{Config: &oauthConfig},
+		SharedSecret:           c.SharedSecret,
+		TrustedUserEmailDomain: c.TrustedUserEmailDomain,
+	}
+	if c.Environment == "test" {
+		authenticator.OAuthClient = auth.FakeOAuthClient{}
+	}
+	return authenticator
+}
+
+func createImageStore(db *sql.DB) store.ImageStore {
+	return store.DBImageStore{DB: db}
+}
+
+func createInstanceStore(db *sql.DB) store.InstanceStore {
+	return store.DBInstanceStore{DB: db}
+}
+
+func createExecutor(c config.Config) exec.Executor {
+	return exec.OSExecutor{DataPath: c.DataPath}
+}
